Remove Docker files, modify requirements and add 'test_schema.yml'
Deleted 'docker-compose.yml' and 'Dockerfile' in the docker directory. Moved 'requirements.txt' from docker directory to backup. Added a 'test_schema.yml' file.

In this commit, 'docker-compose.yml' and 'Dockerfile' have been removed because Docker will no longer be used for deployment. The 'requirements.txt' file was moved to backup because it is not currently needed. A new 'test_schema.yml' file has been added to the root