Update load_config method in GlobalConfig
The load_config method in the GlobalConfig class has been updated to provide additional debugging information and terminate execution upon failure. The path of the configuration file is now resolved and printed out and if there is an exception while loading the file, the program will print an error message and exit.